be/requirements.txt
google-cloud-pubsub==2.18.4
langgraph==0.2.61
pandas==2.2.3
google-cloud-aiplatform==1.73.0
loguru==0.7.2
openpyxl==3.1.5
Crawl4AI>=0.4.247
mysql-connector-python==9.1.0
python-pptx==1.0.2
pdf2image==1.17.0
langfuse==2.57.8
PyPDF2
pytesseract
sentencepiece
pydantic